heal7-project/backend/services/ai-monitoring-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
pyyaml>=6.0.0
aiohttp>=3.8.0
psutil>=5.9.0
prometheus-client>=0.17.0